build: upgrade zone.js (#23108)
PR Close #23108
