+++ BUILD.bazel
@@ -40,12 +40,8 @@ filegroup(
     srcs = [
         "//:node_modules/reflect-metadata/Reflect.js",
         "//:node_modules/zone.js/dist/zone.js",
-        "//:node_modules/zone.js/dist/async-test.js",
-        "//:node_modules/zone.js/dist/sync-test.js",
-        "//:node_modules/zone.js/dist/fake-async-test.js",
+        "//:node_modules/zone.js/dist/zone-testing.js",
         "//:node_modules/zone.js/dist/task-tracking.js",
-        "//:node_modules/zone.js/dist/proxy.js",
-        "//:node_modules/zone.js/dist/jasmine-patch.js",
     ],
 )
