+++ aio/content/examples/setup/src/quickstart-specs.html
@@ -21,12 +21,7 @@
   <script src="node_modules/jasmine-core/lib/jasmine-core/boot.js"></script>
 
   <script src="node_modules/zone.js/dist/zone.js"></script>
-  <script src="node_modules/zone.js/dist/long-stack-trace-zone.js"></script>
-  <script src="node_modules/zone.js/dist/proxy.js"></script>
-  <script src="node_modules/zone.js/dist/sync-test.js"></script>
-  <script src="node_modules/zone.js/dist/jasmine-patch.js"></script>
-  <script src="node_modules/zone.js/dist/async-test.js"></script>
-  <script src="node_modules/zone.js/dist/fake-async-test.js"></script>
+  <script src="node_modules/zone.js/dist/zone-testing.js"></script>
 
   <!-- #docregion files -->
   <script>
