+++ aio/content/examples/testing/src/main-specs.ts
@@ -6,12 +6,7 @@ declare var jasmine;
 
 import './polyfills';
 
-import 'zone.js/dist/async-test';
-import 'zone.js/dist/fake-async-test';
-import 'zone.js/dist/long-stack-trace-zone';
-import 'zone.js/dist/proxy.js';
-import 'zone.js/dist/sync-test';
-import 'zone.js/dist/jasmine-patch';
+import 'zone.js/dist/zone-testing';
 
 import { getTestBed } from '@angular/core/testing';
 import {
