+++ aio/content/examples/webpack/config/karma-test-shim.js
@@ -5,12 +5,7 @@ require('core-js/es6');
 require('core-js/es7/reflect');
 
 require('zone.js/dist/zone');
-require('zone.js/dist/long-stack-trace-zone');
-require('zone.js/dist/proxy');
-require('zone.js/dist/sync-test');
-require('zone.js/dist/jasmine-patch');
-require('zone.js/dist/async-test');
-require('zone.js/dist/fake-async-test');
+require('zone.js/dist/zone-testing');
 
 var appContext = require.context('../src', true, /\.spec\.ts/);
