+++ aio/scripts/_payload-limits.json
@@ -4,7 +4,7 @@
       "uncompressed": {
         "inline": 1971,
         "main": 565539,
-        "polyfills": 40272,
+        "polyfills": 38514,
         "prettify": 14886
       }
     }
