+++ packages/elements/test/BUILD.bazel
@@ -27,11 +27,7 @@ filegroup(
         "//:node_modules/@webcomponents/custom-elements/src/native-shim.js",
         "//:node_modules/reflect-metadata/Reflect.js",
         "//:node_modules/zone.js/dist/zone.js",
-        "//:node_modules/zone.js/dist/async-test.js",
-        "//:node_modules/zone.js/dist/sync-test.js",
-        "//:node_modules/zone.js/dist/fake-async-test.js",
-        "//:node_modules/zone.js/dist/proxy.js",
-        "//:node_modules/zone.js/dist/jasmine-patch.js",
+        "//:node_modules/zone.js/dist/zone-testing.js",
     ],
 )
